refactor: Button 스타일 컴포넌트만 사용하던 부분 수정
